Rework target resolution
Once again rework target resolution to solve cases where if some flow is
a child of multiple schemas, targets got incorrectly resolved.

Before this commit, target resolution happened with the help of
generated schemas: they contained the code which first tried to resolve
a relative segmentId in a current schema and then, if nothing found,
child schemas were asked to resolve the segmentId. This process was
recursive and child schemas were traversed in a not well determined
order which often resulted in the target segment being found in the
wrong schema which led to the incorrect target path.

This process didn't rely on the activePath.

New process of searching for the target is modified so that

1. It is done by NavigationService
2. Schemas only report info about their nodes, and do not forward calls
recursively down to their children (this is now done by NavigationService )
3. Navigating to child schemas now takes `activePath` into account, so
now having "activePath = appFlow.mainFlow.loginFlow.otp" won't search
for a new "credentials" target in "appFlow.loginFlow" schema, it will
correctly find "appFlow.mainFlow.loginFlow"