Remove explicit Schema parameter
Instead make "schema" a public val in a NodeBuilder.

This simplifies instantiation of NavigationService and removes the need
of caching complex schemas. I.e. previously to avoid doing this:

NavigationService(
  schema = AppFlowSchema(Child1Schema(), Child2Schema()),
  nodeBuilder = AppNodeBuilder(
    schema = AppFlowSchema(Child1Schema(), Child2Schema()),
    ...
  )
)

user had to do this:

val schema = AppFlowSchema(Child1Schema(), Child2Schema()),
NavigationService(
  schema,
  nodeBuilder = AppNodeBuilder(
    schema,
    ...
  )
)

But now this is simplified to

NavigationService(
  nodeBuilder = AppNodeBuilder(
    schema = AppFlowSchema(Child1Schema(), Child2Schema()),
    ...
  )
)